+++ assets/components/widgets/raceResults/raceResults.jsx
@@ -88,7 +88,7 @@ var RaceResultEntry = React.createClass({
 			<div className={cx({'classPosition': true})} style={self.getClassColour(entry.classId)}>Class P{entry.positionClass}.</div>
 			<div className="position">{entry.positionOverall}.</div>
 				<div className="manufacturer">
-					<img src={'/img/manufacturers/'+entry.manufacturerId+'.webp'} />
+					<img src={'/render/'+entry.manufacturerId+'/small/'}/>
 				</div>
 				<div className="name">{UI.fixName(entry.name)}</div>
 				<div className="livery">