+++ public/ui.js
@@ -2868,7 +2868,7 @@ var WidgetManager = React.createClass({
 			{ className: cx(classes), onClick: this.onClickDriver },
 			React.createElement(
 				'span',
-				{ className: 'text' },
+				{ className: 'text', 'data-title': this.props.widget.description },
 				this.props.widget.buttonText
 			)
 		);
@@ -2905,9 +2905,29 @@ var Driver = React.createClass({
 		};
 
 		if (timeDiff > 0 && timeDiff < 1000) {
-			classes['close'] = true;
+			classes['close animated flash infinite'] = true;
 		}
 		return cx(classes);
+	},
+	renderPostion: function (position, classId) {
+		var classColour = "rgba(38, 50, 56, 0.8)";
+
+		if (r3eData.classes[classId] != null && r3eClassColours.classes[classId] != null) {
+			classColour = r3eClassColours.classes[classId].colour;
+		}
+
+		const divStyle = {
+			backgroundColor: classColour,
+			position: "absolute",
+			padding: "10px"
+		};
+
+		return React.createElement(
+			'div',
+			{ className: 'position', style: divStyle },
+			'Class P',
+			position
+		);
 	},
 	render: function () {
 		var self = this;
@@ -2949,11 +2969,7 @@ var Driver = React.createClass({
 					(timeDiff / 1000).toFixed(2),
 					's'
 				) : null,
-				React.createElement(
-					'div',
-					{ className: 'current-position' },
-					driver.scoreInfo.positionOverall
-				)
+				self.renderPostion(driver.scoreInfo.positionClass, driver.classId)
 			)
 		);
 	}
@@ -3112,43 +3128,6 @@ UI.components.Controller = React.createClass({
 			return null;
 		}
 		var session = UI.state.sessionInfo;
-
-		// TODO - Add update checker.
-		// // get the current version number and compare to the local package.json
-		// var showUpdateBanner = false;
-		// let base64Url = 'aHR0cHM6Ly9yYXcuZ2l0aHVidXNlcmNvbnRlbnQuY29tL3RvbWphY2ttYW4vcnJlLXNwZWMtdmVyc2lvbi9tYXN0ZXIvdmVyc2lvbi5qc29u';
-		// fetch(atob(base64Url))
-		// .then(res => res.json())
-		// .then((json) => {
-		//   var version = json.version;
-		//
-		// 	// fetch package.json version
-		// 	console.log(window.location);
-		// 	fetch('dev/package.json')
-		// 	.then(result => result.json())
-		// 	.then((jsonA) => {
-		// 			console.log(jsonA);
-		// 	})
-		// 	.catch(e => {
-		// 		console.log(e);
-		// 	});
-		//
-		//
-		// 	console.log(packageJsonVersion);
-		// 	if (version > packageJsonVersion) {
-		// 		showUpdateBanner = true;
-		// 	}
-		// })
-		// .catch(err => {
-		// 	console.log(err);
-		// 	// not important
-		// });
-		//
-		// { showUpdateBanner ?
-		// <div className="updateAvailable animated bounceInDown delay-3s"><p>A New update is now Available in the <a href='https://forum.sector3studios.com/index.php?threads/r3e-reality-modern-broadcast-overlay.12061/' target="_blank">Sector 3 Forum</a>.</p></div>
-		// :
-		// 	null
-		// }
 
 		return React.createElement(
 			'div',